feat(workers): "crypto" global accessible in Worker scope (#5121)
