+++ cli/tests/subdir/worker_crypto.js
@@ -1,0 +1,3 @@
+onmessage = function () {
+  postMessage(!!self.crypto);
+};
